+++ src/vendors/fortiswitch.rs
@@ -775,6 +775,7 @@ impl SwitchVendor for FortiswitchSwitch {
             port_mirrors: vec![],
             snmp: None,
             management_vlan,
+            ..Default::default()
         })
     }
 
@@ -2033,6 +2034,7 @@ mod tests {
             port_mirrors: vec![],
             snmp: None,
             management_vlan: None,
+            warnings: vec![],
         };
 
         // Desired config: add management VLAN 77
@@ -2067,6 +2069,7 @@ mod tests {
             port_mirrors: vec![],
             snmp: None,
             management_vlan: Some(10),
+            warnings: vec![],
         };
 
         // Desired config: change to management VLAN 88
@@ -2098,6 +2101,7 @@ mod tests {
             port_mirrors: vec![],
             snmp: None,
             management_vlan: Some(60),
+            warnings: vec![],
         };
 
         // Desired config: remove management VLAN (defaults to None)
@@ -2121,6 +2125,7 @@ mod tests {
             port_mirrors: vec![],
             snmp: None,
             management_vlan: Some(25),
+            warnings: vec![],
         };
 
         // Desired config: same management VLAN 25